requirements.txt
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cffi==1.15.1
colorama==0.4.6
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.19
django-captcha-admin==0.2
django-cors-headers==4.1.0
django-extensions==3.2.3
django-pivot==1.9.0
django-recaptcha==1.3.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.1.2
fastjsonschema==2.17.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
jdatetime==4.1.1
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
mssql-django==1.3
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.21.6
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.0
prompt-toolkit==3.0.39
psutil==5.9.5
pycparser==2.21
Pygments==2.15.1
pyodbc==4.0.39
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywinpty==2.0.10
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
sqlparse==0.4.4
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.6.3
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
zipp==3.15.0
